use system vimrc if it has same content of guerilla vimrc

+++ infiltrate.sh
@@ -3,13 +3,21 @@
 
 git submodule update --init
 
-# create backup of system's current vim configurations
-mkdir .bak
-cp $HOME/.vimrc `pwd`/`dirname $0`/.bak/vimrc.bak
-cp $HOME/.vim `pwd`/`dirname $0`/.bak/vim.bak
+# vimrcs
+src_vimrc="`pwd`/`dirname $0`/vimrc"
+dest_vimrc="$HOME/.vimrc"
 
-# deploy `vim-guerilla` to the system
-ln -sfi `pwd`/`dirname $0`/vimrc $HOME/.vimrc
-ln -sfi `pwd`/`dirname $0`/vim $HOME/.vim
+if [[ ! -f "$HOME/.vimrc" || "`cat "$src_vimrc"`" != "`cat "$dest_vimrc"`"]]; then
+  # create backup of system's current vim configurations
+  mkdir .bak
+  mv $HOME/.vimrc `pwd`/`dirname $0`/.bak/vimrc.bak
+  mv $HOME/.vim `pwd`/`dirname $0`/.bak/vim.bak
+
+  # deploy `vim-guerilla` to the system
+  ln -sfi `pwd`/`dirname $0`/vimrc $HOME/.vimrc
+  ln -sfi `pwd`/`dirname $0`/vim $HOME/.vim
+fi
+
+# install plugins
 vim +PluginInstall +qall
 vim +PluginClean! +qall
